backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
aiosqlite
pydantic
passlib[bcrypt]
python-jose
PyJWT
pytest
pydantic
pydantic-settings
pgvector